+++ internal/filesystem/file_watcher.go
@@ -1,4 +1,4 @@
-package fs
+package filesystem
 
 import (
 	"log/slog"
