+++ internal/filesystem/file_watcher_test.go
@@ -1,4 +1,4 @@
-package fs
+package filesystem
 
 import (
 	"errors"
